Refactor(#8): 코드리뷰 반영하여 리팩토링
