+++ src/main/java/com/keodam/keodam_backend/app/user/dto/NicknameRequestDto.java
@@ -1,4 +1,4 @@
-package com.keodam.keodam_backend.app.user.dto.req;
+package com.keodam.keodam_backend.app.user.dto;
 
 public record NicknameRequestDto(String nickname) {
 }
