+++ src/main/java/com/keodam/keodam_backend/app/user/dto/RoleRequestDto.java
@@ -1,4 +1,4 @@
-package com.keodam.keodam_backend.app.user.dto.req;
+package com.keodam.keodam_backend.app.user.dto;
 
 import com.keodam.keodam_backend.app.domain.RoleType;
