+++ src/main/java/com/keodam/keodam_backend/app/user/dto/UserResponseDto.java
@@ -1,4 +1,4 @@
-package com.keodam.keodam_backend.app.user.dto.res;
+package com.keodam.keodam_backend.app.user.dto;
 
 import com.keodam.keodam_backend.app.domain.RoleType;
 import lombok.Builder;
