+++ src/main/java/com/keodam/keodam_backend/global/code/status/ErrorStatus.java
@@ -35,6 +35,17 @@ public enum ErrorStatus implements BaseErrorCode {
     USER_NOT_FOUND(HttpStatus.NOT_FOUND,"USER4001","회원 정보를 찾을 수 없습니다."),
     FAMILY_ALREADY(HttpStatus.BAD_REQUEST,"FAMILY4002", "이미 가입한 가족이 존재합니다."),
 
+    // nickname
+    NICKNAME_SPECIAL_CHAR(HttpStatus.BAD_REQUEST, "NICKNAME4001", "특수문자는 닉네임에 포함될 수 없어요."),
+    NICKNAME_DUPLICATE(HttpStatus.BAD_REQUEST, "NICKNAME4002", "중복된 닉네임이에요."),
+    NICKNAME_ALREADY_CHANGED(HttpStatus.BAD_REQUEST, "NICKNAME4003", "닉네임은 최초 설정 후 한 번만 변경할 수 있어요."),
+    NICKNAME_CONTAINS_NAME(HttpStatus.BAD_REQUEST, "NICKNAME4004", "닉네임에 이름은 포함할 수 없어요."),
+    NICKNAME_CONTAINS_BAD_WORD(HttpStatus.BAD_REQUEST, "NICKNAME4005", "닉네임에 비속어는 포함할 수 없어요."),
+
+    // role type
+    INVALID_ROLE_TYPE(HttpStatus.BAD_REQUEST, "ROLE4001", "잘못된 역할 선택입니다."),
+
+
     //quiz
     QUIZ_NOT_FOUND(HttpStatus.NOT_FOUND,"QUIZ4001","퀴즈를 찾을 수 없습니다.");
 
@@ -61,5 +72,4 @@ public enum ErrorStatus implements BaseErrorCode {
                 .httpStatus(httpStatus)
                 .build();
     }
-
 }
